Modify to divide theme-provider.tsx
