+++ playground/src/context/theme-consumer.tsx
@@ -1,0 +1,12 @@
+import * as React from 'react';
+import ThemeContext from './theme-context';
+
+interface ThemedButtonProps {}
+
+export default function ToggleThemeButton(props: ThemedButtonProps) {
+  return (
+    <ThemeContext.Consumer>
+      {({ theme, toggleTheme }) => <button style={theme} onClick={toggleTheme} {...props} />}
+    </ThemeContext.Consumer>
+  );
+}
